requirements.txt
google-cloud-storage
gunicorn
flask
